Introducing a level below
This isn't finished yet. This commit has multiple elements:

For the interpreter, we use ';' instead of 'z' to indicate a pause.
This frees the 'z' key to act as map swap key. When swapping the map,
we exchange the current map with the map below. In order to facilitate
this, I split the initialization code into initialization and
resetting the map.