Fix wall mode
Removing walls was broken in Wall Mode. This is fixed. In addition to
that, when in Wall Mode, well make an extra effort to find the wall
we're on: (x, y) where the angle could be 0° or 90°, (x-1, y) where
the angle could be 180° and (x, y-1) where the angle could be 270°.