Traps, pillars and statues no longer advance
In a link, traps, pillars and statues no longer advance the position.
The benefit is that these three kinds of tiles can now have variants.
The README file now has instructions on how to add your own tile (a
well) as a variant of a similar looking tile (the statue).

Sadly, this is also an incompatible change.