Prepare Download link on Chrome
Instead of just hiding the Prepare Download link, Chrome will hide all
the remaining text elements and their children due to some sort of bug.
If we move the How to Save link to the front, it works.

I noticed that the Prepare Link code doesn't really work all that well.
I think I'm going to abandon this branch.