Bidirectional communication
On the server side, clients are now given numbers that are independent
of their position in the array of clients as that would result in an
error if a client left the map and rejoined it. The position of the
other clients in the array would change and thus their numbers were
wrong.

The server is sending messages in the format Ctrl-A id Ctrl-B command
Ctrl-C. This vaguely inspired by ASCII control codes (beginning of
heading, beginning of text, end of text).

On the client side, we no longe link to the Gridmap server. We simply
provide a Javascript link to join maps. When interpreting commands
sent from remote clients, we remember the position of their respective
pens. Visually, this isn't quite as pleasing.