Don't move(null)
The move(null) happened if interpretCont was called with the advance
flag set, even if the direction was null. This call was unnecessary
and it caused Pen.ts to be null because of how it was set. This in
turn resulted in the code "w.w" to be messed up.