Fix deleting a character from the astral plane
When adding deleting a character such as 🍄 from a label, we run into
a problem: Javascript exposes the two surrogate halves that form a
single astral symbol. Using backspace on such a label gives us illegal
garbage. So now we're testing the string and then we remove one or two
characters from the end. And handle the undo case, too.