Do not update position with Shift key
It used to be impossible to use Gridmapper with just the mouse: in
order to pick something from the menu on the side you had to move
there with the mouse and that moved the pointer, too. Now you can
prevent the pointer from following the mouse by holding the Shift key.