Small improvements to saving and loading
SVG, PNG and TXT now save immediately (unless you're on IE), no
questions asked. This improves the situation when clicking on the
links: SVG and PNG files no longer open in the browser after
downloading. The trade off is that using a right click and picking
Save As… from the context menu no longer works.