Make sure grottos are adjusted when drawing floor
This is done by calling checkRockyNeighbors whenever a tile is drawn
on the floor, and by that I mean that draw calls functions which at
the end will all call replace. After running the replace command, we
run checkRockyNeighbors, which will maybe call replaceColoredFloor
which again calls replace. In order to prevent this from running amok,
we pass along a recursive flag. If this flag is set, replace will
*not* run checkRockyNeighbors (because we already are).