Fix handling of " and '
Previously, if you used "abc" to insert small text, it worked as
intended, but as soon as you used 'abc' to insert large text, "abc"
would no longer work as intended: it would also insert large text.
This commit makes sure to remove the class attribute if the label type
is null.